created separate components for login and register inputs; refactored loginregister
