+++ src/components/LoginRegister.js
@@ -1,38 +1,22 @@
 import React, { useState } from "react";
+import { LoginInputs } from "./LoginInputs";
+import { RegisterInputs } from "./RegisterInputs";
 import {
   useRouteMatch,
+  useHistory,
   Link as ReactLink,
   Switch,
   Route,
 } from "react-router-dom";
-import { useForm } from "react-hook-form";
 import axios from "axios";
 // styles
-import {
-  Flex,
-  Text,
-  Link,
-  FormErrorMessage,
-  FormLabel,
-  FormControl,
-  Button,
-  Input,
-  InputGroup,
-  InputRightElement,
-} from "@chakra-ui/core";
+import { Flex, Text, Link } from "@chakra-ui/core";
 
 function LoginRegister({ user, setUser, storedToken }) {
-  let { path, url } = useRouteMatch();
+  let { path } = useRouteMatch();
+  let history = useHistory();
 
-  // form state
-  const [showPassword, setShowPassword] = useState(false);
-  const [showRetype, setShowRetype] = useState(false);
   const [regError, setRegError] = useState("");
-  const { register, handleSubmit, errors, formState, getValues } = useForm();
-
-  // form functions
-  const handlePassClick = () => setShowPassword((bool) => !bool);
-  const handleRetypeClick = () => setShowRetype((bool) => !bool);
 
   // user login/register functions
   const onRegisterSubmit = (data) => {
@@ -53,52 +37,21 @@ function LoginRegister({ user, setUser, storedToken }) {
           ...user,
           name: res.username,
         }));
+        history.push("/auth/login");
       })
       .catch((err) => {
         setRegError("There was an error registering, try again.");
       });
   };
 
-  // validation functions
-  function validateUsername(value) {
-    let error;
-    if (!value) {
-      error = "Username is required";
-    } else if (value.length > 15 || value.length < 2) {
-      error = "Username must be 2 - 15 characters long";
-    }
-    return error || true;
-  }
-  function validatePassword(value) {
-    let error;
-    const expression = new RegExp(
-      "^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$%^&*-]).{0,}$"
-    );
-    if (!value) {
-      error = "Password is required";
-    } else if (value.length < 8) {
-      error = "Password must be at least 8 characters long";
-    } else if (expression.test(value) === false) {
-      error =
-        "Password must include a capital letter, number, and special character";
-    }
-    return error || true;
-  }
-  function validateRetypePassword(value) {
-    let error;
-    const top_password = getValues("password");
-    if (!value) {
-      error = "Must re-type your password";
-    } else if (value !== top_password) {
-      error = "Passwords must match";
-    }
-    return error || true;
-  }
+  const onLoginSubmit = (data) => {
+    console.log(data);
+  };
 
   return (
     <Flex direction="column" m="0 auto">
       <Flex
-        width="100%"
+        width="400px"
         mx="auto"
         mb="2rem"
         justify="space-evenly"
@@ -135,71 +88,17 @@ function LoginRegister({ user, setUser, storedToken }) {
             Login or Register to save your cupcakes :)
           </Text>
         </Route>
+        {/* REGISTER */}
         <Route path={`${path}/register`}>
-          {/* REGISTER */}
-          <form onSubmit={handleSubmit(onRegisterSubmit)}>
-            <FormControl isRequired isInvalid={errors.username}>
-              {/* USERNAME */}
-              <FormLabel htmlFor="username">Username</FormLabel>
-              <Input
-                type="text"
-                placeholder="Username"
-                name="username"
-                ref={register({ validate: validateUsername })}
-              />
-              <FormErrorMessage>
-                {errors.username && errors.username.message}
-              </FormErrorMessage>
-            </FormControl>
-            {/* PASSWORD */}
-            <FormControl isRequired isInvalid={errors.password}>
-              <FormLabel htmlFor="password">Password</FormLabel>
-              <InputGroup size="md">
-                <Input
-                  pr="3rem"
-                  type={showPassword ? "text" : "password"}
-                  placeholder="Enter Password"
-                  name="password"
-                  ref={register({ validate: validatePassword })}
-                />
-                <InputRightElement width="4rem">
-                  <Button h="1.75rem" size="sm" onClick={handlePassClick}>
-                    {showPassword ? "Hide" : "Show"}
-                  </Button>
-                </InputRightElement>
-              </InputGroup>
-              <FormErrorMessage>
-                {errors.password && errors.password.message}
-              </FormErrorMessage>
-            </FormControl>
-            {/* RETYPE PASSWORD */}
-            <FormControl isRequired isInvalid={errors.retypepassword}>
-              <FormLabel htmlFor="retypepassword">Re-type Password</FormLabel>
-              <InputGroup size="md">
-                <Input
-                  pr="3rem"
-                  type={showRetype ? "text" : "password"}
-                  placeholder="Re-type Password"
-                  name="retypepassword"
-                  ref={register({ validate: validateRetypePassword })}
-                />
-                <InputRightElement width="4rem">
-                  <Button h="1.75rem" size="sm" onClick={handleRetypeClick}>
-                    {showRetype ? "Hide" : "Show"}
-                  </Button>
-                </InputRightElement>
-              </InputGroup>
-              <FormErrorMessage>
-                {errors.retypepassword && errors.retypepassword.message}
-              </FormErrorMessage>
-            </FormControl>
-            <Button isLoading={formState.isSubmitting} type="submit">
-              Submit
-            </Button>
-            {regError && <Text color="red">{regError}</Text>}
-          </form>
+          <RegisterInputs
+            onRegisterSubmit={onRegisterSubmit}
+            regError={regError}
+          />
         </Route>
-        <Route path={`${path}/login`}>Login</Route>
+        {/* LOGIN */}
+        <Route path={`${path}/login`}>
+          <LoginInputs onLoginSubmit={onLoginSubmit} />
+        </Route>
       </Switch>
     </Flex>
   );
